Fixed some compile-time errors on 64-bit architecture.  Specifically, pointers are now cast to intptr_t instead of int, and uintptr_t instead of unsigned (these types are defined in <inttypes.h>).  Also, to get a Boolean value that is true just when a pointer is nonzero, we now use (ptr != NULL) rather than trying to cast the pointer to BOOL.

+++ c_pgmtk/src/include/cart/inlines.h
@@ -3,6 +3,7 @@
 
 #include <math.h>
 #include <ctype.h>
+#include <inttypes.h>
 #include "datadefs.h"
 
 #ifdef WIN32
@@ -180,7 +181,7 @@ CART_INLINE float icxSetSign(float f, float sign)
 CART_INLINE void icxAssertAligned( void* ptr , const int align = CART_MEM_ALIGN)
 { 
 	assert(ptr);
-	assert ( ( ((int)ptr) & (align-1) ) == 0 );
+	assert ( ( ((intptr_t)ptr) & (align-1) ) == 0 );
 }
 
 CART_INLINE void* icxAlignedMalloc(int size , const int align = CART_MEM_ALIGN)
@@ -287,7 +288,7 @@ CART_INLINE int icxGetVarNumStates(int type )
 { return (type & CX_CLASSIFIER_VAR_TYPE_NUM_STATES_MASK); }
 
 CART_INLINE BOOL icxIsChunkLoaded(CxClassifierSampleChunk* chunk )
-{ return (int)(chunk) && (chunk)->is_loaded; }
+{ return (intptr_t)(chunk) && (chunk)->is_loaded; }
 
 CART_INLINE float icxGetSampleWeight(CxClassifierSampleChunk* chunk, int sample_idx )
 { return *(float *)((char *)(chunk)->weights_data + (sample_idx) * (chunk)->weights_step); }
